backend.wakatsuki/requirements.txt
mistralai==1.9.10
python-dotenv==1.1.1
